swapped Build and Cure
